+++ sdk/core/core-rest-pipeline/src/index.ts
@@ -71,4 +71,9 @@ export {
   AuthorizeRequestOptions,
   AuthorizeRequestOnChallengeOptions
 } from "./policies/bearerTokenAuthenticationPolicy";
+export {
+  armChallengeAuthenticationPolicy,
+  ARMChallengeAuthenticationPolicyOptions,
+  ARMChallengeAuthenticationPolicyName
+} from "./policies/armChallengeAuthenticationPolicy";
 export { ndJsonPolicy, ndJsonPolicyName } from "./policies/ndJsonPolicy";